adding examples
