+++ src/amy-example-rp2040.c
@@ -8,6 +8,7 @@
 #include <math.h>
 
 #include "amy.h"
+#include "examples.h"
 #if PICO_ON_DEVICE
 
 #include "hardware/clocks.h"
@@ -93,101 +94,6 @@ struct audio_buffer_pool *init_audio() {
     return producer_pool;
 }
 
-// Emulate the Tulip "drums()" example via event calls.
-void drums(int64_t start, int loops) {
-    //int64_t start = amy_sysclock();
-    struct event e = amy_default_event();
-    e.time = start;
-
-    float volume = 0.2;
-
-    int oscs[] = {0, 2, 3, 4, 5, 6};
-    int patches[] = {1, 5, 0, 10, 10, 5};
-    e.wave = PCM;
-    e.freq = 0;
-    e.velocity = 0;
-    for (int i = 0; i < sizeof(oscs) / sizeof(int); ++i) {
-        e.osc = oscs[i];
-        e.patch = patches[i];
-        amy_add_event(e);
-    }
-    // Update high cowbell.
-    e = amy_default_event();
-    e.time = start;
-    e.osc = 5;
-    e.midi_note = 70;
-    amy_add_event(e);
-
-    // osc 7 : bass
-    e = amy_default_event();
-    e.time = start;
-    e.osc = 7;
-    e.wave = SAW_DOWN;
-    e.filter_freq = 2500.0;
-    e.resonance = 5.0;
-    e.filter_type = FILTER_LPF;
-    e.breakpoint_target[0] = TARGET_AMP + TARGET_FILTER_FREQ;
-    amy_add_event(e);
-
-    char bp0msg[] = "v0A0,1,500,0.2,25,0\0";
-    bp0msg[1] = '0' + e.osc;
-    amy_play_message(bp0msg);
-
-    const int bass = 1 << 0;
-    const int snare = 1 << 1;
-    const int hat = 1 << 2;
-    const int cow = 1 << 3;
-    const int hicow = 1 << 4;
-
-    int pattern[] = {bass+hat, hat+hicow, bass+hat+snare, hat+cow, hat, hat+bass, snare+hat, hat};
-    int bassline[] = {50, 0, 0, 0, 50, 52, 51, 0};
-
-    e = amy_default_event();
-    e.time = start;
-    while (loops--) {
-        for (int i = 0; i < sizeof(pattern) / sizeof(int); ++i) {
-            e.time += 250;
-
-            int x = pattern[i];
-            if(x & bass) {
-                e.osc = 0;
-                e.velocity = 4.0 * volume;
-                amy_add_event(e);
-            }
-            if(x & snare) {
-                e.osc = 2;
-                e.velocity = 1.5 * volume;
-                amy_add_event(e);
-            }
-            if(x & hat) {
-                e.osc = 3;
-                e.velocity = 1 * volume;
-                amy_add_event(e);
-            }
-            if(x & cow) {
-                e.osc = 4;
-                e.velocity = 1 * volume;
-                amy_add_event(e);
-            }
-            if(x & hicow) {
-                e.osc = 5;
-                e.velocity = 1 * volume;
-                amy_add_event(e);
-            }
-
-            e.osc = 7;
-            if(bassline[i]>0) {
-                e.velocity = 0.5 * volume;
-                e.midi_note = bassline[i] - 12;
-            } else {
-                e.velocity = 0;
-            }
-            amy_add_event(e);
-            e.midi_note = -1;
-        }
-    }
-}
-
 
 void core1_main() {
     while(1) {
@@ -241,40 +147,12 @@ int main() {
         sleep_ms(250);
     }
 
-#if AMY_HAS_REVERB == 1
-    config_reverb(0, REVERB_DEFAULT_LIVENESS, REVERB_DEFAULT_DAMPING, REVERB_DEFAULT_XOVER_HZ);
-#endif
-
-#if AMY_HAS_CHORUS == 1
-    config_chorus(0.8, CHORUS_DEFAULT_MAX_DELAY);
-#endif
+    example_reverb();
+    example_chorus();
 
     struct audio_buffer_pool *ap = init_audio();
-
-    // Play a few notes in FM
-    struct event e = amy_default_event();
     int64_t start = amy_sysclock();
-    int osc_inc;
-    e.time = start;
-    e.osc = 0;
-    e.wave = ALGO;
-    e.patch = 20;
-    osc_inc = 9;
-    int notes[] = {60, 70, 64, 68, 72, 82, 76, 80, 74, 78, 80, 58};
-    e.velocity = 0.2;
-
-    if (true) {
-        for (int i = 0; i < sizeof(notes) / sizeof(int); ++i) {
-            e.midi_note = notes[i];
-            e.pan = 0.5 + 0.5 * ((2 * (i %2)) - 1);
-            e.patch++;
-            amy_add_event(e);
-            e.osc += osc_inc;
-            e.time += 1000;
-        }
-    } else {
-        drums(e.time + 500, 10 /* loops */);
-    }
+    example_multiimbral_fm(start);
 
     for (int i = 0; i < 5000; ++i) {
         rp2040_fill_audio_buffer(ap);
