+++ src/amy-example.c
@@ -3,6 +3,7 @@
 
 
 #include "amy.h"
+#include "examples.h"
 #include "libminiaudio-audio.h"
 
 int main(int argc, char ** argv) {
@@ -34,42 +35,15 @@ int main(int argc, char ** argv) {
                 break; 
         } 
     }
-
+    int64_t start =amy_sysclock();
     amy_start();
     amy_live_start();
     amy_reset_oscs();
 
-    // Play a few notes in FM
-    #if AMY_HAS_REVERB == 1
-    config_reverb(2, REVERB_DEFAULT_LIVENESS, REVERB_DEFAULT_DAMPING, REVERB_DEFAULT_XOVER_HZ); 
-    #endif
+    example_reverb();
+    example_chorus();
+    example_multimbral_fm(start);
 
-    #if AMY_HAS_CHORUS == 1
-    config_chorus(0.8, CHORUS_DEFAULT_MAX_DELAY);
-    #endif
-    
-    struct event e = amy_default_event();
-    int64_t start = amy_sysclock();
-    e.time = start;
-
-    e.velocity = 1;
-    e.wave = KS;
-    e.feedback = F2S(0.996f);
-    e.patch = 15;
-    e.osc = 0;
-    e.midi_note = 60;
-    amy_add_event(e);
-    /*
-    e.time = start + 500;
-    e.osc += 9; // remember that an FM patch takes up 9 oscillators
-    e.midi_note = 64;
-    amy_add_event(e);
-    
-    e.time = start + 1000;
-    e.osc += 9;
-    e.midi_note = 68;
-    amy_add_event(e);
-    */
     // Now just spin for 5s
     while(amy_sysclock() - start < 5000) {
         usleep(THREAD_USLEEP);
