+++ src/examples.h
@@ -1,0 +1,13 @@
+// examples.h
+// examples.c
+// sound examples
+
+#include "amy.h"
+
+
+void example_reverb() ;
+void example_chorus() ;
+void example_ks(int64_t start) ;
+void example_fm(int64_t start) ;
+void example_multimbral_fm(int64_t start) ;
+void example_drums(int64_t start, int loops);
